Add Nginx to the develop compose file
This change set adds a thrid container to the develope compose file, this
container runs an nginx instance working as a proxy server which proxies
the docker frontend via a suffix (/docker-ui). This will help to make sure
that our links are relative and work when being proxied.

Currently the nginx config is setup to work with the wan the app is currently
done up, which required some extra locations to handle the app json file,

ideally we'll update the app to require a simpilar nginx configuration