Move from Bower to NPM
This change set removes :fire: Bower as the dependancy package manager and
switches over to NPM. Bower requiers all packages to be in git and pulls
them out of a public git repo. NPM on the other hand is a registry of
packages and it has become the go to for JS package dependancy for both
Node and Browser applications. The Bower website even suggests moving
off Bower in favor of NPM/Yarn.