Avoid crash if fail to fetch docker catalog
If the query to the Docker registry to fetch all the images fails the service
was crashing because it tried to parse the body of a 500 or 404 as JSON but it
is returned as plain text.

This change set updates the query to check the status of the response and if
it is not a 200 then set the return to null (i.e. empty array and undefines)
so that the rest of the code will run correctly and avoid exceptions being
thrown.