:bug:Fix app-mode.json not working through proxy pass
If you have a proxy in play which adds a path to the URL the app was not
correctly fetching the app-mode.json and app-version.json as the query for them
was not respecting the base path.

The code which fetch these json files was using an absolute path (i.e. they
started with a /) and thus ignored the base path. By removing the leading
slash they became relative paths and thus used the base path.

With this change you can have your proxy re-write the base path to match
the path used by the proxy and thus the query to fetch the JSON files
will use the correct absolute path was resolved.

This fixes issue 126