Use Maintainer label for the author
Docker has deprecated the  stanza in favour of the
label (https://docs.docker.com/v17.03/engine/reference/builder/)

This change set updates the UI to first look for the maintainer label when
showing the author of the image. If the Image does not have a mainatiner
label it will fall back tot he maintainer stanza.

This was pulled from PR #174 by 0xC4N1 in the parent repo