:bug:Fix behaviour of switching reposPerPage
By default the repos list shows 10 repos per page but the user can select to
show more. The links to show more repos simply set a different value to the
route (i.e. repositories/20) however this drops the values for the pagation.

The pagation works by setting the repo/image of the next page at the end of the path
this repo/image value was dropped when the user updated how many repos to show
on each page thus the user was sent back to the first page.

This change set updates the repos per page selection links to perserve the
pagation values.

This was taken from the parent repository's PR #131 submitted by xiaodot