:art:Clean up the next link parser
Per the TODO _Can we clean this up a bit?_ this change set cleans up
the code used to parse out the last value via the next link header.

In the current implementation we use substring, indexOf, and replace
which results in a bit of nesting and unclear code. This change set
updates that code to use Regular Expressions and named functions to
make the code more clear.

Tests were added to ensure hat this refactor didn't break anything