:bug:Fix Next button disable when no namespace
The Next button is only enabled if it has both a last namespace and last
repo but in cases where there is no namespace the Next button was disabled.

This change set pulls the logic for enabeling/disabeling the next button (and
previous button) from the View and into the controller. It then sets a flag
to disable the next (and prviouse)  button.

The repository service which use to parse the Link header and return both
the last repository (name) and last namespace has been updated to simply
return the 'last repository (name)' abstracting away the namespace. This
was done because we always use the two values together and having them
seperated ment we would have to always check if one didn't exist.

Now the next (previous) button is correctly enabled/disabled regardless
if the next link contains a namespace or not.