:green_heart:Add tests for tag sorting
This change set adds tests for the new tag sorting by date feature.

It changed the syntax for the query of tags a bit to use the .then()
api vs. passing the handler function as a second parameter to the query()
api as that just made writing the test easier and more _promisy_