:bug:Fix bug were descending sort only works after asc sort
This change set fixs a bug where if the array has yet to be sorted and the sort
direction is desc the array is simply reversed. If the array was already in
desc order this would result in the array now being sorted asc which is not
the desired behaviour.

This change set applies the sort regardless of direction to ensure the list is
always properly sorted.