:fire:Remove main and directory properties from package
This change set removes the main and directory properties from the package
json as this package json does not publish a module which can be installed
by another application; this packge is an endpoint and thus has no main
as the main will be the compailed angularJS app. It will also never be
pushed to the registry as again it is an endpoint and thus does not
need to specify which directires should be included in the package.