Update the Node/NPM version
The package.json specifies which version of Node this package will run on
and which version of NPM is needed to install this package. This package
won't be published but it halpes when using NPM to install packages
used in the building and testing of the application.

Thus this change set update sthe version to Node 8 and the latest verison
of NPM used with Node 8