:shirt:Switch JSHint to ESLint
ESlint has loads more rules and support then JSHint. This change set
switches the project over to ESlint and includes the Airbnb style guide.

The Airbnb JS Style guid is a collection of well documented JS ESLint Rules
which help prevent simple bugs and improves code readability.

This change is using the Legacy rules (i.e. pre ES2015) as we don't have
babel in the project so we must still to pre-ES2015 for browser compatability