:shirt:Switch to modern  ESLint rules
When I switched from JSHint to ESLint I setup the application to use the
legacy (pre-ES2015) rules because we didn't have Babel thus needed to use
the legacy syntax/feature set to remain compatible on older browsers.

Now that the build pipeline includes Babel we can write the application using
the latest language features while still _compiling_ an application which will
work with older browsers. Because of that this change set switches the linter
rules to the modern (post-ES2015) rules.